adding patch files to git so they won't get lost

+++ src/main/scala/org/intellij/scala/bundle/package.scala
@@ -33,9 +33,7 @@ package object bundle {
       input.close()
     } catch {
       case _: Throwable =>
-        if (!destination.delete()) {
-          destination.deleteOnExit()
-        }
+        if (!destination.delete()) destination.deleteOnExit()
     }
   }
 
@@ -48,9 +46,7 @@ package object bundle {
       input.close()
     } catch {
       case _: Throwable =>
-        if (!to.delete()) {
-          to.deleteOnExit()
-        }
+        if (!to.delete()) to.deleteOnExit()
     }
   }
